+++ lib/kitchen/directions/bake_chapter_introductions/v1.rb
@@ -2,6 +2,9 @@
 
 module Kitchen::Directions::BakeChapterIntroductions
   class V1
+    # <b>DEPRECATED:</b> Please use <tt>v2</tt> instead.
+    warn '[DEPRECATION] `v1` is deprecated.  Please use `v2` instead.'
+
     def bake(book:)
       book.chapters.each do |chapter|
         introduction_page = chapter.introduction_page
@@ -17,10 +20,18 @@ module Kitchen::Directions::BakeChapterIntroductions
           "> :not([data-type='metadata']):not(.splash):not(.has-splash)"
         ).cut
 
-        chapter_objectives_html =
-          Kitchen::Directions::BakeChapterIntroductions.bake_chapter_objectives(
-            chapter: chapter
-          )
+        chapter_objectives_html = chapter.non_introduction_pages.map do |page|
+          <<~HTML
+            <div class="os-chapter-objective">
+              <a class="os-chapter-objective" href="##{page.title[:id]}">
+                <span class="os-number">#{chapter.count_in(:book)}.#{page.count_in(:chapter)}</span>
+                <span class="os-divider"> </span>
+                <span data-type="" itemprop="" class="os-text">#{page.title.children[0].text}</span>
+              </a>
+            </div>
+          HTML
+        end.join('')
+
         chapter_outline =
           Kitchen::Directions::BakeChapterIntroductions.bake_chapter_outline(
             chapter_objectives_html: chapter_objectives_html